Improved seperation of concerns for theme-provider
