+++ packages/views/src/components/theme-provider/hooks/use-theme-store.ts
@@ -1,0 +1,16 @@
+import { create } from 'zustand'
+import { persist } from 'zustand/middleware'
+
+import { FullTheme, ThemeState } from '../types'
+
+export const useThemeStore = create<ThemeState>()(
+  persist(
+    set => ({
+      theme: 'system-std-standard', // Default theme
+      setTheme: (newTheme: FullTheme) => set({ theme: newTheme })
+    }),
+    {
+      name: 'canary-ui-theme' // LocalStorage key
+    }
+  )
+)
