+++ packages/views/src/components/theme-provider/hooks/use-theme.ts
@@ -1,0 +1,8 @@
+import { useThemeStore } from './use-theme-store'
+
+export const useTheme = () => {
+  const theme = useThemeStore(state => state.theme)
+  const setTheme = useThemeStore(state => state.setTheme)
+
+  return { theme, setTheme }
+}
